+++ src/controllers/network/index.ts
@@ -133,14 +133,12 @@ export const networkGiverCommand: Command = {
             name: "address",
             title: "Giver address",
             type: "string",
-            defaultValue: "",
         },
         {
             name: "signer",
             alias: "s",
             title: "Signer to be used with giver",
             type: "string",
-            defaultValue: "",
         },
         {
             name: "value",
@@ -184,7 +182,6 @@ export const networkCredsCommand: Command = {
     title: "Set credentials for network authentication",
     args: [
         nameArg,
-
         {
             name: "project",
             alias: "p",