+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### New
 
 - Include network giver account details in `network info <network>` command.
+- Giver's signer must be specified explicitly (default network signer will not be used).
+- Include EverNode SE Default Giver Keys to signers repository.
 
 ### Fixed